+++ CHANGELOG.md
@@ -10,6 +10,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1944 Rename "Intended Use" in the footer as "Product labelling"
 ### Engineering use
 - YLP-1624 Fix error messages when inviting a patient already in the team
+- YLP-1625 Display an explicit error when trying to share data with another patient
 - YLP-1680 getPatientData Api method has a wrong parameter
 - YLP-1726 Rework signup with Auth0
 - YLP-1805 Push error to the backend